+++ crates/bevy_plugin/src/assets.rs
@@ -1,27 +1,18 @@
 use crate::prelude::*;
-use bevy::asset::{FileAssetIo, LoadedAsset};
+use bevy::asset::LoadedAsset;
 use bevy::prelude::*;
 use bevy::{
     asset::{AssetLoader, LoadContext},
     utils::BoxedFuture,
 };
-use std::path::{Path, PathBuf};
 
 #[derive(Debug, Default)]
 pub(crate) struct YarnSlingerAssetLoaderPlugin;
 
 impl Plugin for YarnSlingerAssetLoaderPlugin {
     fn build(&self, app: &mut App) {
-        app.register_type::<Localizations>()
-            .register_type::<Localization>()
-            .register_type::<FileGenerationMode>()
-            .add_asset::<YarnFile>()
-            .init_asset_loader::<YarnFileAssetLoader>()
-            .add_system(
-                generate_missing_line_ids_in_yarn_file
-                    .pipe(panic_on_err)
-                    .run_if(is_in_development),
-            );
+        app.add_asset::<YarnFile>()
+            .init_asset_loader::<YarnFileAssetLoader>();
     }
 }
 
@@ -46,46 +37,6 @@ impl AssetLoader for YarnFileAssetLoader {
     }
 }
 
-fn generate_missing_line_ids_in_yarn_file(
-    mut events: EventReader<AssetEvent<YarnFile>>,
-    mut assets: ResMut<Assets<YarnFile>>,
-    asset_server: Res<AssetServer>,
-) -> SystemResult {
-    for event in events.iter() {
-        if let AssetEvent::Created { handle } = event {
-            let file = assets.get_mut(handle).unwrap();
-            let Some(source_with_added_ids) = add_tags_to_lines(file.clone())? else {
-                // File already contains all line IDs.
-                return Ok(())
-            };
-
-            // If this fails, the asset was created at runtime and doesn't exist on disk.
-            if let Some(asset_path) = asset_server.get_handle_path(handle.clone()) {
-                let assets_path = get_assets_dir_name(&asset_server)?;
-                let path_within_asset_dir: PathBuf =
-                    [assets_path.as_ref(), asset_path.path()].iter().collect();
-
-                std::fs::write(&path_within_asset_dir, &source_with_added_ids)
-                    .context(
-                        format!("Failed to overwrite Yarn file at {} with new line IDs. \
-                                 Aborting because localization requires all lines to have IDs, but this file is missing some.",
-                                path_within_asset_dir.display()))?;
-            }
-            file.source = source_with_added_ids;
-        }
-    }
-    Ok(())
-}
-
-fn get_assets_dir_name(asset_server: &AssetServer) -> Result<impl AsRef<Path> + '_> {
-    let asset_io = asset_server.asset_io();
-    let file_asset_io = asset_io.downcast_ref::<FileAssetIo>().context(
-        "Failed to downcast asset server IO to `FileAssetIo`. \
-    The vanilla Bevy `FileAssetIo` is the only one supported by Yarn Slinger",
-    )?;
-    Ok(file_asset_io.root_path())
-}
-
 fn read_yarn_file<'a>(
     bytes: &'a [u8],
     load_context: &'a mut LoadContext,
@@ -99,18 +50,4 @@ fn read_yarn_file<'a>(
         .context("Yarn file name is not valid UTF-8")?
         .to_owned();
     Ok(YarnFile { file_name, source })
-}
-
-/// Adapted from <https://github.com/YarnSpinnerTool/YarnSpinner-Console/blob/main/src/YarnSpinner.Console/Commands/TagCommand.cs#L11>
-fn add_tags_to_lines(yarn_file: YarnFile) -> YarnCompilerResult<Option<String>> {
-    let existing_tags = YarnCompiler::new()
-        .with_compilation_type(CompilationType::StringsOnly)
-        .add_file(yarn_file.clone())
-        .compile()
-        .unwrap()
-        .string_table
-        .into_iter()
-        .filter_map(|(key, string_info)| (!string_info.is_implicit_tag).then_some(key))
-        .collect();
-    YarnCompiler::add_tags_to_lines(yarn_file.source, existing_tags)
 }
