+++ crates/bevy_plugin/src/localization/config.rs
